Rename get container to get instance
